Sun, Jul 30, 2023, 2:15 PM -05:00

+++ finance/app.py
@@ -75,7 +75,7 @@ def buy():
         stock = lookup(symbol.upper())
         if stock == None:
             return apology("ERROR: Symbol does not exist. Please input.")
-        if int(shares) < 0:
+        if float(shares) < 0:
              return apology("ERROR: Amount of shares not allowed. Please input correctly.")
         transaction_value = shares * stock["price"]
 
@@ -219,7 +219,7 @@ def sell():
         if shares < 0:
              return apology("ERROR: Amount of shares not allowed. Please input correctly.")
 
-        transaction_value = int(shares) * int(stock["price"])
+        transaction_value = shares * stock["price"]
 
         user_id = session["user_id"]
         user_cash_db = db.execute("SELECT cash FROM users WHERE id = :id", id=user_id)
